refactor: fixing circular dependencies, adding apiError action
