+++ src/api/api.reducer.ts
@@ -11,3 +11,5 @@ export const apiRequest = createAction(
   `${name} apiRequest`,
   (meta: ApiMeta, payload?: any): ApiRequestResponse => ({ payload, meta }),
 )
+
+export const apiError = createAction<any>(`${name} apiError`)
